+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Added pagination header to Timeline API responses.
+- Use `name` and `summary` attributes to create post title.
 
 ### Changed